Updated source_utils to get dependencies

+++ source_utils.bash
@@ -10,11 +10,23 @@
 # Copyright (c) 2019 Denis Semenenko
 ###########################################################################
 
+#------< Constants >------#
 PREFIX='[Dolsem Bash Utils]'
 REPO_URL='https://github.com/dolsem/bash-utils'
 BASE_URL="${REPO_URL}/raw/master"
 CACHE_DIR="$(dirname $0)/.bash-utils"
 
+#------< Dependency graph (tree) >------#
+deps_assert=()
+deps_filesystem=()
+deps_network=(os)
+deps_os=()
+deps_prompt=(filesystem assert)
+deps_string=(os)
+deps_term=()
+deps_validation=()
+
+#------< Helpers >------#
 clone_repo() {
   echo "${PREFIX} Getting all utilities..."
   if [[ -z $noprogress ]]; then
@@ -27,32 +39,45 @@ clone_repo() {
   fi
 }
 
-download_one() {
-  echo "${PREFIX} Getting $1 utilities..."
+get_one() {
   mkdir -p "$CACHE_DIR"
   if cd "$CACHE_DIR"; then
-    if [[ -z $noprogress ]]; then
-      wget -O "$1.bash" "${BASE_URL}/$1.bash" -q --show-progress --progress=bar:noscroll
-    else
-      wget -O "$1.bash" "${BASE_URL}/$1.bash" -q
+    if [[ ! -f $1.bash ]]; then
+      echo "${PREFIX} Getting $1 utilities..."
+      if [[ -z $noprogress ]]; then
+        wget -O "$1.bash" "${BASE_URL}/$1.bash" -q --show-progress --progress=bar:noscroll
+      else
+        wget -O "$1.bash" "${BASE_URL}/$1.bash" -q
+      fi
+      cd - 1>/dev/null
     fi
-    cd - 1>/dev/null
+  else
+    echo
+    echo "${PREFIX} $CACHE_DIR: access denied"
+    exit 1
   fi
   if [[ ! $? -eq 0 ]]; then
     exit $?
   fi
+  echo -e '\033[1A\033[52C Done.'
 }
 
+get_dependencies() {
+  eval "local deps=(\${deps_$1[@]})"
+  if [ ! ${#deps[@]} -eq 0 ]; then
+    echo "${PREFIX} Getting $1 utilities dependencies..."
+    for dep_util in "${deps[@]}"; do
+      get_one $dep_util
+    done
+  fi
+}
+
+#------< Main >------#
 if [[ -z $util ]]; then
   clone_repo
   source "${CACHE_DIR}/*.bash"
 else
-  download_one $util
+  get_one $util
+  get_dependencies $util
   source "${CACHE_DIR}/${util}.bash"
-fi
-
-if [[ $? -eq 0 ]]; then
-  echo "${PREFIX} Done."
-else
-  exit $?
 fi
